\name{HPO.db-internal}
\title{Internal HPO.db objects}
\alias{HP}
\alias{HPO_DB_AnnDbBimap_seeds}
\alias{createAnnObjs.HPO_DB}
\alias{datacache}
\alias{HP,HPOTerms-method}
\alias{HP,HPOTermsAnnDbBimap-method}
\alias{HP,character-method}
\alias{as.list,HPOTermsAnnDbBimap-method}
\description{Internal HPO.db objects.}
\details{These are not to be called by the user.}
\keyword{internal}