\name{HPO.db}
\alias{HPO.db}
\alias{HPO}
\title{Human Phenotype Ontology Annotation Data Package }
\description{

  Welcome to the HPO.db annotation Package.  The purpose of this package
  is to provide detailed information about the latest version of the
  Phenotype Ontology.  
  
  You can learn what objects this package supports with the following command:
  
  \code{ls("package:HPO.db")} 
  
  Each of these objects has their own manual page detailing where
  relevant data was obtained along with some examples of how to use it.
}

\examples{
  ls("package:HPO.db")
}
\keyword{datasets}











