\name{HPOTermsAnnDbBimap}
\docType{class}

\alias{HPOTermsAnnDbBimap}
\alias{class:HPOTermsAnnDbBimap}
\alias{HPOTermsAnnDbBimap-class}

\title{Class "HPOTermsAnnDbBimap" }

\description{
  A sub-class of  Bimap, specific for HPO.db. Please see \code{\link[AnnotationDbi:Bimap]{Bimap}} for details.
}


\seealso{
   \code{\link[AnnotationDbi:Bimap]{Bimap}},
   \code{\link{HPOTerms}}
}

\examples{
   class(HPOTERM)
}

\keyword{classes}
\keyword{interface}












